pyproject.toml
[tool.poetry]
name = "smcat"
version = "0.1.0"
description = "Sitemap cat for catting sitemaps."
authors = ["datadavev <605409+datadavev@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Scrapy = "^2.5.0"
ciso8601 = "^2.2.0"
requests = "^2.26.0"
asyncio = "^3.4.3"
hashids = "^1.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = "^5.5"
pytest-twisted = "^1.13.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
